pyproject.toml
[tool.poetry]
authors = ["Etuloser <1494136313@qq.com>"]
description = ""
name = "e-flask"
version = "0.1.0"

[tool.poetry.dependencies]
python = "^3.6"
flask = "^1.1.2"
python-dotenv = "^0.15.0"
flask-sqlalchemy = "^2.4.4"
xlrd = "^2.0.1"
xlwt = "^1.3.0"
pymysql = "^1.0.2"
kafka-python = "^2.0.2"
flask-restful = "^0.3.8"
flask-apscheduler = "^1.11.0"

[tool.poetry.dev-dependencies]
autopep8 = "^1.5.5"
pylint = "^2.7.0"

[build-system]
build-backend = "poetry.core.masonry.api"
requires = ["poetry-core>=1.0.0"]

[[tool.poetry.source]]
default = true
name = "tsinghua"
url = "https://pypi.tuna.tsinghua.edu.cn/simple"
